+++ artemis-core/artemis/src/main/java/com/artemis/EntityEdit.java
@@ -36,7 +36,6 @@ public final class EntityEdit {
 	 * @param componentKlazz Class to create.
 	 * @return Newly instanced component.
 	 */
-	// FIXME: BREAKING CHANGFE - no longe replaces component (desirable)
 	public <T extends Component> T create(Class<T> componentKlazz) {
 		return cm.getMapper(componentKlazz).create(entityId);
 	}